Refactored the UI to break the content apart from the display of the content
