+++ src/Components/PageComponents/Page.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
-import { Grid, Grow } from '@material-ui/core';
+import Grid from '@material-ui/core/Grid';
 
 import Fade from '@material-ui/core/Fade';
