+++ src/Content/SectionInformation/ProjectsInformation.js
@@ -1,0 +1,21 @@
+import TextTwistLogo from '../../Content/Resources/TextTwistBackground.png';
+import ASORTDLogo from '../../Content/Resources/ASORTD.png';
+
+export default {
+  projects: [
+    {
+      name: 'Text Twist',
+      description: 'A clone of the game Text-Twist. Developed using React, Bootstrap, and PHP',
+      codeLink: 'https://github.com/scottbenton/TextTwist-UI',
+      deployedLink: 'https://scottbenton.github.io/TextTwist-UI/',
+      image: TextTwistLogo,
+    },
+    {
+      name: 'ASORTD',
+      description: 'Algorithm SORTing Display. A website that animates sorting algorithms, and allows users to input their own sorting algorithms to animate.',
+      codeLink: 'https://github.com/willswire/asortd',
+      deployedLink: 'https://asortd.info',
+      image: ASORTDLogo,
+    },
+  ]
+}
