+++ src/Content/SectionInformation/WorkInformation.js
@@ -1,0 +1,41 @@
+import WSFSLogo from '../../Content/Resources/WSFSLogo.png';
+import CSELogo from '../../Content/Resources/SportsEmporiumLogo.png';
+import CVLogo from '../../Content/Resources/CVLogo.png';
+
+export default {
+  workplaces: [
+    {
+      name: 'WSFS Bank',
+      dates: 'June 2019-Present',
+      jobTitle: 'Software Developer Intern',
+      notes: [
+        "Used Agile methodologies to bring a piece of internal bank software from start to finish",
+        "Developed a user friendly Material Design website using React.",
+        "Learned React"
+      ],
+      logo: WSFSLogo,
+    },
+    {
+      name: 'Carlisle Sports Emporium',
+      dates: 'December 2015 - April 2019',
+      jobTitle: 'Roamer',
+      notes: [
+        "Ensured safety of customers and equipment on attractions",
+        "Developed problem solving skills by performing maintenance on a variety of arcade machines and other equipment",
+        "Adapted quickly to new roles and expectations in the job"
+      ],
+      logo: CSELogo,
+    },
+    {
+      name: 'Cumberland Valley Marching Band',
+      dates: 'June - August 2016, 2018',
+      jobTitle: 'Volunteer Educator',
+      notes: [
+        "Taught high school and middle school students music and marching skills",
+        "Guided student leaders towards effective leadership methods",
+        "Developed educational skills by teaching new concepts to students"
+      ],
+      logo: CVLogo,
+    },
+  ]
+}
